Allow one-way compression (#18)
* Allow one-way compression

A client may negotiate for compression over a connection but send uncompressed
frames. In such cases, the server is expected to skip decompression of incoming
frames but compress the outgoing frames. This change checks the rsv1 bit of
every incoming frame and invokes the decompressor only if the rsv1 bit is set,
i.e if the frame is compressed.

* autobahn: upgrade pip

* Temporarily disabling autobahn on xenial-5.0.1