Add automatic stale connection cleanup (#40)
* Add automatic stale connection cleanup

Motivation:

On Kitura-WebSocket https://github.com/IBM-Swift/Kitura-WebSocket/pull/48/
introduces an automatic connection cleanup feature. Using this,
`WebSocketService`s can configure an optional timeout value. If a connection
remains idle for half of the configured timeout value, the server sends a
heartbeat ping to the client. In a following time period equal to
half of the configured timeout value, if a pong is received on the connection,
the latter is kept alive. If not, the connection is closed.

In `Kitura-WebSocket-NIO`, we implement this using the `IdleStateHandler` which
is configured with half of the configured timeout value. When we receive the
first `IdleStateEvent`, we send a heartbeat ping to the client. If a pong is
received before the next `IdleStateEvent`, the connection is kept alive. If the
next `IdleStateEvent` is received before the pong, the connection is closed.

This commit also includes a new client framework for testing. It is difficult
to write the connection cleanup tests using the existing adhoc WebSocket client
framework implemented in `KituraTest.swift` and `KituraTest+Frames.swift`. The
new client framework is called `WebSocketClient`. It uses a callback model and
is largely work-in-progress. We must move all of the existing tests to the new
framework in due time.

* Log error and close channel on unknown error

* Propagate event if not IdleStateEvent